+++ Nico/nicolasburtz-web/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "my-app",
+  "name": "nicolasburtz-web",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "react": "^17.0.2",
