+++ Nico/nicolasburtz-web/public/manifest.json
