+++ Nico/nicolasburtz-web/src/App.js
@@ -1,0 +1,21 @@
+import logo from './logo.svg';
+import './css/App.css';
+
+function App() {
+
+  return (
+
+    <div className="App">
+
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+
+      </header>
+      
+    </div>
+
+  );
+
+}
+
+export default App;
