+++ Nico/nicolasburtz-web/src/css/App.css
@@ -9,7 +9,7 @@
 
 @media (prefers-reduced-motion: no-preference) {
   .App-logo {
-    animation: App-logo-spin infinite 20s linear;
+    animation: 0.5s linear 1s infinite alternate slidein; 
   }
 }
